Add support for starting the process to be profiled
- Docs update
- The usual double dash command line "syntax" is used for this feature
- Parsing the command line in this case is especially hard to do correctly, so an admittedly subpar solution was chosen utilizing a regex
- The process is started in a suspended state, and resumed after the profiler is started
- Introduced a means to launch processes in suspended state
- Processes can now be terminated
- Minor enum flags bugfix
- Minor refactor of "collecting" process information (PID, name) of targets
- Introduced both command line and profile tests for the new functionality

+++ Sources/etwprof/Profiler/ETLReloggerProfiler.cpp
@@ -173,7 +173,7 @@ bool ETLReloggerProfiler::EnableProvider (const IETWBasedProfiler::ProviderInfo&
     return true;
 }
 
-uint16_t ETLReloggerProfiler::GetNumberOfProfiledProcesses ()
+uint32_t ETLReloggerProfiler::GetNumberOfProfiledProcesses ()
 {
     return 1;
 }
